formatting

+++ aphrodite/common/gguf.py
@@ -8,7 +8,6 @@ from typing import Any, Literal, NamedTuple, TypeVar, Union
 import numpy as np
 import numpy.typing as npt
 
-
 GGUF_MAGIC = 0x46554747  # "GGUF"
 GGUF_VERSION = 3
 GGUF_DEFAULT_ALIGNMENT = 32
@@ -16,8 +15,8 @@ READER_SUPPORTED_VERSIONS = [2, GGUF_VERSION]
 
 
 class GGMLQuantizationType(IntEnum):
-    F32  = 0
-    F16  = 1
+    F32 = 0
+    F16 = 1
     Q4_0 = 2
     Q4_1 = 3
     Q5_0 = 6
@@ -31,20 +30,20 @@ class GGMLQuantizationType(IntEnum):
     Q6_K = 14
     Q8_K = 15
     IQ2_XXS = 16
-    IQ2_XS  = 17
+    IQ2_XS = 17
     IQ3_XXS = 18
-    IQ1_S   = 19
-    IQ4_NL  = 20
-    IQ3_S   = 21
-    IQ2_S   = 22
-    IQ4_XS  = 23
+    IQ1_S = 19
+    IQ4_NL = 20
+    IQ3_S = 21
+    IQ2_S = 22
+    IQ4_XS = 23
 
 
 QK_K = 256
 # Items here are (block size, type size)
 GGML_QUANT_SIZES = {
-    GGMLQuantizationType.F32:  (1, 4),
-    GGMLQuantizationType.F16:  (1, 2),
+    GGMLQuantizationType.F32: (1, 4),
+    GGMLQuantizationType.F16: (1, 2),
     GGMLQuantizationType.Q4_0: (32, 2 + 16),
     GGMLQuantizationType.Q4_1: (32, 2 + 2 + 16),
     GGMLQuantizationType.Q5_0: (32, 2 + 4 + 16),
@@ -62,24 +61,26 @@ GGML_QUANT_SIZES = {
     GGMLQuantizationType.IQ3_XXS: (256, 2 + 3 * QK_K // 8),
     GGMLQuantizationType.IQ1_S: (256, 2 + QK_K // 8 + QK_K // 16),
     GGMLQuantizationType.IQ4_NL: (32, 2 + 32 // 2),
-    GGMLQuantizationType.IQ3_S: (256, 2 + QK_K // 4 + QK_K // 32 + QK_K // 8 + QK_K // 64),
+    GGMLQuantizationType.IQ3_S:
+    (256, 2 + QK_K // 4 + QK_K // 32 + QK_K // 8 + QK_K // 64),
     GGMLQuantizationType.IQ2_S: (256, 2 + QK_K // 4 + QK_K // 32 + QK_K // 32),
     GGMLQuantizationType.IQ4_XS: (256, 2 + 2 + QK_K // 64 + QK_K // 2),
 }
 
+
 class GGUFValueType(IntEnum):
-    UINT8   = 0
-    INT8    = 1
-    UINT16  = 2
-    INT16   = 3
-    UINT32  = 4
-    INT32   = 5
+    UINT8 = 0
+    INT8 = 1
+    UINT16 = 2
+    INT16 = 3
+    UINT32 = 4
+    INT32 = 5
     FLOAT32 = 6
-    BOOL    = 7
-    STRING  = 8
-    ARRAY   = 9
-    UINT64  = 10
-    INT64   = 11
+    BOOL = 7
+    STRING = 8
+    ARRAY = 9
+    UINT64 = 10
+    INT64 = 11
     FLOAT64 = 12
 
     @staticmethod
@@ -133,23 +134,25 @@ class GGUFReader:
 
     # Note: Internal helper, API may change.
     gguf_scalar_to_np: dict[GGUFValueType, type[np.generic]] = {
-        GGUFValueType.UINT8:   np.uint8,
-        GGUFValueType.INT8:    np.int8,
-        GGUFValueType.UINT16:  np.uint16,
-        GGUFValueType.INT16:   np.int16,
-        GGUFValueType.UINT32:  np.uint32,
-        GGUFValueType.INT32:   np.int32,
+        GGUFValueType.UINT8: np.uint8,
+        GGUFValueType.INT8: np.int8,
+        GGUFValueType.UINT16: np.uint16,
+        GGUFValueType.INT16: np.int16,
+        GGUFValueType.UINT32: np.uint32,
+        GGUFValueType.INT32: np.int32,
         GGUFValueType.FLOAT32: np.float32,
-        GGUFValueType.UINT64:  np.uint64,
-        GGUFValueType.INT64:   np.int64,
+        GGUFValueType.UINT64: np.uint64,
+        GGUFValueType.INT64: np.int64,
         GGUFValueType.FLOAT64: np.float64,
-        GGUFValueType.BOOL:    np.bool_,
+        GGUFValueType.BOOL: np.bool_,
     }
 
-    def __init__(self, path: os.PathLike[str] | str, mode: Literal['r' | 'r+' | 'c'] = 'r'):
-        self.data = np.memmap(path, mode = mode)
+    def __init__(self,
+                 path: os.PathLike[str] | str,
+                 mode: Literal['r' | 'r+' | 'c'] = 'r'):
+        self.data = np.memmap(path, mode=mode)
         offs = 0
-        if self._get(offs, np.uint32, override_order = '<')[0] != GGUF_MAGIC:
+        if self._get(offs, np.uint32, override_order='<')[0] != GGUF_MAGIC:
             raise ValueError('GGUF magic invalid')
         offs += 4
         temp_version = self._get(offs, np.uint32)
@@ -160,13 +163,21 @@ class GGUFReader:
             temp_version = temp_version.newbyteorder(self.byte_order)
         version = temp_version[0]
         if version not in READER_SUPPORTED_VERSIONS:
-            raise ValueError(f'Sorry, file appears to be version {version} which we cannot handle')
+            raise ValueError(
+                f'Sorry, file appears to be version {version} which we cannot handle'
+            )
         self.fields: OrderedDict[str, ReaderField] = OrderedDict()
         self.tensors: list[ReaderTensor] = []
-        offs += self._push_field(ReaderField(offs, 'GGUF.version', [temp_version], [0], [GGUFValueType.UINT32]))
+        offs += self._push_field(
+            ReaderField(offs, 'GGUF.version', [temp_version], [0],
+                        [GGUFValueType.UINT32]))
         temp_counts = self._get(offs, np.uint64, 2)
-        offs += self._push_field(ReaderField(offs, 'GGUF.tensor_count', [temp_counts[:1]], [0], [GGUFValueType.UINT64]))
-        offs += self._push_field(ReaderField(offs, 'GGUF.kv_count', [temp_counts[1:]], [0], [GGUFValueType.UINT64]))
+        offs += self._push_field(
+            ReaderField(offs, 'GGUF.tensor_count', [temp_counts[:1]], [0],
+                        [GGUFValueType.UINT64]))
+        offs += self._push_field(
+            ReaderField(offs, 'GGUF.kv_count', [temp_counts[1:]], [0],
+                        [GGUFValueType.UINT64]))
         tensor_count, kv_count = temp_counts
         offs = self._build_fields(offs, kv_count)
         offs, tensors_fields = self._build_tensors_fields(offs, tensor_count)
@@ -180,7 +191,7 @@ class GGUFReader:
             offs += self.alignment - padding
         self._build_tensors(offs, tensors_fields)
 
-    _DT = TypeVar('_DT', bound = npt.DTypeLike)
+    _DT = TypeVar('_DT', bound=npt.DTypeLike)
 
     # Fetch a key/value metadata field by key.
     def get_field(self, key: str) -> Union[ReaderField, None]:
@@ -191,29 +202,37 @@ class GGUFReader:
         return self.tensors[idx]
 
     def _get(
-        self, offset: int, dtype: npt.DTypeLike, count: int = 1, override_order: None | Literal['I' | 'S' | '<'] = None,
+        self,
+        offset: int,
+        dtype: npt.DTypeLike,
+        count: int = 1,
+        override_order: None | Literal['I' | 'S' | '<'] = None,
     ) -> npt.NDArray[Any]:
         count = int(count)
-        itemsize = int(np.empty([], dtype = dtype).itemsize)
+        itemsize = int(np.empty([], dtype=dtype).itemsize)
         end_offs = offset + itemsize * count
-        return (
-            self.data[offset:end_offs]
-            .view(dtype = dtype)[:count]
-            .newbyteorder(override_order or self.byte_order)
-        )
+        return (self.data[offset:end_offs].view(
+            dtype=dtype)[:count].newbyteorder(override_order
+                                              or self.byte_order))
 
     def _push_field(self, field: ReaderField, skip_sum: bool = False) -> int:
         if field.name in self.fields:
-            raise KeyError(f'Duplicate {field.name} already in list at offset {field.offset}')
+            raise KeyError(
+                f'Duplicate {field.name} already in list at offset {field.offset}'
+            )
         self.fields[field.name] = field
         return 0 if skip_sum else sum(int(part.nbytes) for part in field.parts)
 
-    def _get_str(self, offset: int) -> tuple[npt.NDArray[np.uint64], npt.NDArray[np.uint8]]:
+    def _get_str(
+            self, offset: int
+    ) -> tuple[npt.NDArray[np.uint64], npt.NDArray[np.uint8]]:
         slen = self._get(offset, np.uint64)
         return slen, self._get(offset + 8, np.uint8, slen[0])
 
     def _get_field_parts(
-        self, orig_offs: int, raw_type: int,
+        self,
+        orig_offs: int,
+        raw_type: int,
     ) -> tuple[int, list[npt.NDArray[Any]], list[int], list[GGUFValueType]]:
         offs = orig_offs
         types: list[GGUFValueType] = []
@@ -238,7 +257,8 @@ class GGUFReader:
             aparts: list[npt.NDArray[Any]] = [raw_itype, alen]
             data_idxs: list[int] = []
             for idx in range(alen[0]):
-                curr_size, curr_parts, curr_idxs, curr_types = self._get_field_parts(offs, raw_itype[0])
+                curr_size, curr_parts, curr_idxs, curr_types = self._get_field_parts(
+                    offs, raw_itype[0])
                 if idx == 0:
                     types += curr_types
                 idxs_offs = len(aparts)
@@ -263,7 +283,7 @@ class GGUFReader:
         offs += int(offset_tensor.nbytes)
         return ReaderField(
             orig_offs,
-            str(bytes(name_data), encoding = 'utf-8'),
+            str(bytes(name_data), encoding='utf-8'),
             [name_len, name_data, n_dims, dims, raw_dtype, offset_tensor],
             [1, 3, 4, 5],
         )
@@ -277,19 +297,22 @@ class GGUFReader:
             offs += int(raw_kv_type.nbytes)
             parts: list[npt.NDArray[Any]] = [kv_klen, kv_kdata, raw_kv_type]
             idxs_offs = len(parts)
-            field_size, field_parts, field_idxs, field_types = self._get_field_parts(offs, raw_kv_type[0])
+            field_size, field_parts, field_idxs, field_types = self._get_field_parts(
+                offs, raw_kv_type[0])
             parts += field_parts
             self._push_field(ReaderField(
                 orig_offs,
-                str(bytes(kv_kdata), encoding = 'utf-8'),
+                str(bytes(kv_kdata), encoding='utf-8'),
                 parts,
                 [idx + idxs_offs for idx in field_idxs],
                 field_types,
-            ), skip_sum = True)
+            ),
+                             skip_sum=True)
             offs += field_size
         return offs
 
-    def _build_tensors_fields(self, offs: int, count: int) -> tuple[int, list[ReaderField]]:
+    def _build_tensors_fields(self, offs: int,
+                              count: int) -> tuple[int, list[ReaderField]]:
         tensor_fields = []
         for _ in range(count):
             field = self._get_tensor(offs)
@@ -297,9 +320,11 @@ class GGUFReader:
             tensor_fields.append(field)
         return offs, tensor_fields
 
-    def _build_tensors(self, start_offs: int, fields: list[ReaderField]) -> None:
+    def _build_tensors(self, start_offs: int,
+                       fields: list[ReaderField]) -> None:
         tensors = []
         for field in fields:
+            # pylint: disable=unused-variable
             _name_len, name_data, _n_dims, dims, raw_dtype, offset_tensor = field.parts
             ggml_type = GGMLQuantizationType(raw_dtype[0])
             n_elems = np.prod(dims)
@@ -316,15 +341,15 @@ class GGUFReader:
             else:
                 item_count = n_bytes
                 item_type = np.uint8
-            tensors.append(ReaderTensor(
-                name = str(bytes(name_data), encoding = 'utf-8'),
-                tensor_type = ggml_type,
-                shape = dims,
-                n_elements = n_elems,
-                n_bytes = n_bytes,
-                data_offset = data_offs,
-                data = self._get(data_offs, item_type, item_count),
-                field = field,
-            ))
+            tensors.append(
+                ReaderTensor(
+                    name=str(bytes(name_data), encoding='utf-8'),
+                    tensor_type=ggml_type,
+                    shape=dims,
+                    n_elements=n_elems,
+                    n_bytes=n_bytes,
+                    data_offset=data_offs,
+                    data=self._get(data_offs, item_type, item_count),
+                    field=field,
+                ))
         self.tensors = tensors
-
